+++ tests/test_wikipron/test_extract.py
@@ -4,7 +4,6 @@ import requests_html
 from wikipron.extract import EXTRACTION_FUNCTIONS
 from wikipron.extract.core import _skip_pron
 from wikipron.extract.default import extract_word_pron_default
-from . import config_factory
 
 
 @pytest.mark.parametrize(
@@ -30,11 +29,10 @@ def test_extraction_functions_have_the_same_signature(func):
         # GH-105: Dashed prons are skipped.
         ("ɑb-", "eng", True, True),
         # Spaces in Chinese prons are not skipped.
-        ("ɕjɛ tu", "cmn", False, True),
+        ("ɕjɛ tu", "cmn", False, False),
         # Non-breaking spaces are not skipped.
-        ("zinda ɡi", "per", True, True),
+        ("zinda ɡi", "per", False, False),
     ],
 )
 def test__skip_pron(pron, iso639_key, skip_spaces, expected):
-    config = config_factory(key=iso639_key, skip_spaces_pron=skip_spaces)
-    assert _skip_pron(pron, config) == expected
+    assert _skip_pron(pron, skip_spaces) == expected